requirements.txt
asgiref==3.7.2
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
cryptography==42.0.5
Django==5.0.2
django-autoslug==1.9.9
django-ckeditor-5==0.2.12
django-crispy-forms==2.1
django-filter==23.5
django-js-asset==2.2.0
django-taggit==5.0.1
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
izitoast==0.0.2
Markdown==3.4.1
pillow==10.2.0
progressbar2==4.4.1
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyparsing==3.1.1
python-utils==3.8.2
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
